WIP: Compile with the C standard library
BRANCH=none
BUG=none
TEST=none

+++ common/util.c
@@ -63,14 +63,14 @@ int parse_bool(const char *s, int *dest)
 {
 	/* off, disable, false, no */
 	if (!strcasecmp(s, "off") || !strncasecmp(s, "dis", 3) ||
-	    tolower(*s) == 'f' || tolower(*s) == 'n') {
+	    tolower((int)*s) == 'f' || tolower((int)*s) == 'n') {
 		*dest = 0;
 		return 1;
 	}
 
 	/* on, enable, true, yes */
 	if (!strcasecmp(s, "on") || !strncasecmp(s, "ena", 3) ||
-	    tolower(*s) == 't' || tolower(*s) == 'y') {
+	    tolower((int)*s) == 't' || tolower((int)*s) == 'y') {
 		*dest = 1;
 		return 1;
 	}
